+++ src/other_serial/C/spmat_utils.h
@@ -167,7 +167,7 @@ sparsemat_t *    random_graph(int64_t n, graph_model model, edge_type edge_type,
 sparsemat_t *    random_sparse_matrix(int64_t nrows, int64_t ncols, double density, int values, int64_t seed);
 
 sparsemat_t *    read_matrix_mm(char * name);
-void             resolve_edge_prob_and_nz_per_row(double * edge_prob, int64_t * nz_per_row,
+void             resolve_edge_prob_and_nz_per_row(double * edge_prob, double * nz_per_row,
                                                   int64_t numrows, self_loops loops);
 int64_t          sort_nonzeros( sparsemat_t *mat);
 void             spmat_stats(sparsemat_t *mat);
